backend/requirements.txt
asgiref==3.4.1
black==21.9b0
chardet==4.0.0
click==8.0.3
cycler==0.10.0
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
multimetric==1.3.0
mypy-extensions==0.4.3
mysqlclient==2.0.3
numpy==1.21.3
pandas==1.3.4
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.3
python-dateutil==2.8.2
pytz==2021.3
read==0.0.2
regex==2021.10.23
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tomli==1.2.2
typing-extensions==3.10.0.2
